implemented workaround for some Graph API responses not returning Content-Type on success.
This fix implements a workaround for certain responses from the Graph API (for example the Create Team call) where the API does not return the Content-Type header, even though documentation says it should return with said error.

A bug ticket has been raised to microsoft about this issue.

work around works by wrapping the RuntimeException used to indicate that the Content-Type header is missing with a custom NoContentTypeHeaderException and including the ResponseInterface $response in this wrapper exception.
This exception is backwards compatible with RuntimeException so existing catches of this exception will continue to work as they did before.